refactor: Add BestModelOne/Two and BestResultsOne/Two classes
This way, plotting results is easier (there is no need to pass the data again to the plotting functions),
 and there is an OOP structure where objects lend themselves to certain actions.

========================

Change draw_zero:bool to ref_val:Optional[float] in best.plot.
This matches the ArviZ API (apart from the possibility of None).

========================

Change variable_name to var_name in best.plot.
This matches the ArviZ API closer (apart from the possibility of a list).